refactor: Тестируемый класс вынесен в отдельный пакет
